+++ source/rok4_tasks/rok4_tasks/manager_based/locomotion/velocity/mdp/__init__.py
@@ -6,3 +6,4 @@ from .actions import *  # noqa: F401, F403
 from .commands import *  # noqa: F401, F403
 from .observations import *  # noqa: F401, F403
 from .rewards import *  # noqa: F401, F403
+from .symmetry import *  # noqa: F401, F403